refactor server bootstrap resources
